consolidate utils related to entry points
