+++ packages/core/src/plugins/vite/internal-package-resolution.ts
@@ -4,8 +4,9 @@ import type { Plugin } from 'vite';
 import { normalizePath } from 'vite';
 
 import type { EnhancedConfig } from '../../config';
-import { getPackageEntryPoints, getPackages } from '../../utils/get-packages';
-import type { Packages } from '../../utils/get-packages';
+import type { Packages } from '../../utils/entry-points';
+import { getPackages } from '../../utils/entry-points';
+import { getPackageEntryPoints } from '../../utils/entry-points';
 
 export const internalPackageResolution = (config: EnhancedConfig): Plugin => {
   let packages: Packages | undefined;
@@ -41,7 +42,7 @@ export const internalPackageResolution = (config: EnhancedConfig): Plugin => {
         if (!packageEntries.get(packageName)) {
           packageEntries.set(
             packageName,
-            (await getPackageEntryPoints({ packageRoot: root })).map(
+            (await getPackageEntryPoints(root)).map(
               ({ entryName }) => entryName,
             ),
           );
